validator/pyproject.toml
[project]
name = "validator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "azure-functions>=2.1.0",
    "azure-identity>=1.25.3",
    "jsonschema>=4.26.0",
    "msgraph-sdk>=1.56.0",
    "pytest>=9.0.3",
    "simpleeval>=1.0.7",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
